% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_job_script}
\alias{create_job_script}
\title{Creation of the bash file containing the instruction for the array job to
send to che cluster.}
\usage{
create_job_script(.num_jobs, .cmd_file)
}
\arguments{
\item{.num_jobs}{Number of jobs to be sent to the cluster. Every job is a fcs
file to e processed.}

\item{.cmd_file}{Name of the file containing the commands to filter the data.
It must correspond to the number of lines in \emph{.num_jobs}.}
}
\value{
The function writes the file \emph{job.sh} and it doesn't return any
  value..
}
\description{
This function creates the bash file \emph{job.sh} containing the instructions
to be sent to the cluster in order to filter the data based on their
scattering profile. The bash file contains the number of datasets to be
processed, that is why we cannot have a universal job.sh to be called for any
analysis, but we have to create it every time. The log of the jobs are stored
in the folder 'qlog'.
}

