% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{propagate_index_info}
\alias{propagate_index_info}
\title{Add info to the preprocessed statistics.}
\usage{
propagate_index_info(.pls, .info)
}
\arguments{
\item{.pls}{a tibble containing the preprocessed statistics.}

\item{.info}{a tibble containing the column \emph{dir} and the information to
be added. The matching between the information in \emph{.info} and the
datasets in \emph{.pls} is done according to the directory names.}
}
\value{
A tibble containing the original statistics and the added
  information.
}
\description{
Takes information sotred in a tibble \emph{.info} and adds them to the tibble
of preprocessed statistics. The incorporation is done by matching the
folders.
}

