% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary.R
\name{scattering_filter}
\alias{scattering_filter}
\title{Filtering of FCM files based on their scattering profile.}
\usage{
scattering_filter(.dirs, .data2preproc, .f_par, .filter_preproc_namer, .force,
  .preproc_script, .jobs_cmd_name, .verbose)
}
\arguments{
\item{.dirs}{List of directories containing the files to be preprocessed.}

\item{.data2preproc}{Function that transforms the directory name of the FCM
files to the directory name where to store preprocessed data.}

\item{.f_par}{List containing the parameters used to preprocess the data. It
must contain the elements \enumerate{ \item \emph{channels}: the names of
the channels containing the signals to be analysed, in the order: forward
scattering first signal, forward scattering second signal, side scattering
first signal, side scattering second signal, fluorescence. \item
\emph{scattering_threshold}: a cell is considered viable if its probability
to come from the main scattering distribution is bigger or equal to this
threshold. \item \emph{scattering_num_cells}: retain only the
\emph{scattering_num_cells} with the highest poterior distribution; this is
a way to specify a specific number of cells to retain from the filtering. If
\emph{scattering_threshold} is specified, \emph{scattering_num_cells} will
be ignored. \item \emph{file_pattern}: the pattern fo the FCM files to be
analysed. Useful if the directories contain files that must not be analysed.
\item \emph{delta_shot_noise}: the amplitude of the shot noise. \item
\emph{autofluo_wells}: Tibble containing for every dir the names of the
empty plasmids wells }}

\item{.filter_preproc_namer}{Function that takes the directory name where to
store the preprocessed files and returns the name and path of a Rdata file
where to store the preprocessed data of the single files. By default, if
\emph{path} is the directory of the preprocessed data and the file is called
abc.fcs, then the preprocessed data for that file is stored in
\emph{path/abc.Rdata}.}

\item{.preproc_script}{Name of the script that perform the filtering of the
scattering signals and the lognormal fit of the fluorescence. Deafult to
\emph{preproc.R}.}

\item{.jobs_cmd_name}{Name of the file where to store the list of commands to
send to the cluster. Default to \emph{filtering_cmd}}

\item{.verbose}{Should some information about the several stages the function
goes through be displayed? Default to FALSE}
}
\value{
The function doesn't return any value, but it saves the statistics of
 the filtered data in Rdata files in the folder \code{.data2preproc} with the
 name given by \code{.filter_preproc_name}.
}
\description{
\code{scattering_filter} filters all the FCM files in a set of folders based
on their scattering profile and it fits the survived cells with a mixture of
lognormal and uniform distribution. It stores the final result for each FCM
file in the directory specified by \code{.data2preproc}, such that they can be
loaded for subsequent analysis (e.g. by \code{remove_noise}).
}
\details{
This function loops over all the dirs provided and check for FCM files that
have not yet been filtered. The filtered information are stored in a path
given by the function .data2preproc and with a filename given by
.filter_preproc_namer. More specificaly, what the function does is to check
for every FCM files if the file storing the filtered info is present. If it is
not present, or if .force is TRUE, it adds the command to filter the file in
the file CMD in the working directory. At the end, it creates a shell job file
to execute an array job which is sent to the cluster. IMPORTANT: dirnames must
be given as absolute paths or as relative the current working directory.
}

